\name{version-class}
\docType{class}
\alias{version-class}


\title{Class "version"}
\description{	Contains version information of the ARF package}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("version", ...)}.

}
\section{Slots}{
	 \describe{
    \item{\code{version}:}{Version number. }
    \item{\code{build}:}{Build number. }
    \item{\code{update}:}{Update number. }
  }
}
\section{Methods}{
  \describe{
    \item{slots:}{Slots can be accessed via the standard accessor and replacment methods.}

	 }
}

\author{ Wouter D. Weeda }

\keyword{classes}
