\name{writeData}
\alias{writeData}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ write data in NIFTI/ANALYZE format }
\description{
  writes a vector of fMRI data in a NIFTI/ANALYZE file
}
\usage{
writeData(headinf, datavec)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{headinf}{ An object of class 'nifti.header' }
  \item{datavec}{ A numeric vector containing data. }
}
\details{
 
}
\value{
  returns TRUE if succeeded.
}
\references{  }
\author{ Wouter D. Weeda }
\note{ }
\seealso{  }
\examples{

}

\keyword{methods}
