\name{writeHeader}
\alias{writeHeader}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ write NIFTI/ANALYZE header information to a file }
\description{
  write NIFTI/ANALYZE header information to a file
}
\usage{
writeHeader(headinf)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{headinf}{ An object of class 'nifti.header' }
}

\value{
  returns TRUE if succeeded.
}

\author{ Wouter D. Weeda }


\keyword{ methods }

