\name{writeHeader}
\alias{writeHeader}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ write NIFTI/ANALYZE header information to a file }
\description{
  write NIFTI/ANALYZE header information to a file
}
\usage{
writeHeader(headinf)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{headinf}{ An object of class 'nifti.header' }
}
\details{
  
}
\value{
  returns TRUE if succeeded.
}
\references{  }
\author{ Wouter D. Weeda }
\note{ }
\seealso{  }
\examples{
 
}

\keyword{ methods }

